% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_boxplot}
\alias{circ_boxplot}
\title{Draw boxplot}
\usage{
circ_boxplot(x, label = NA, height = 1, cex = 0.7, col = grey(0.95),
  probs = seq(0, 1, 0.25), ...)
}
\description{
Functions calculates medians using the \code{quantile} function
from the \code{circular} package.
}
\examples{
# simple boxplot
x <- c(0.06, 0.1, 0.19, 0.9, 0.1, 0.1, 0.07)
circ_plot()
circ_par_init()
ring_set(1)
circ_points(x, .5 , col="red", pch=16)
ring_set(2)
circ_boxplot(x)
}

