% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_convert}
\alias{circ_convert}
\title{Workhorse: original range to cartesian coords according to \code{circ_par} settings.}
\usage{
circ_convert(x, y, to.cart = TRUE)
}
\arguments{
\item{x,y}{Coordinates from original data range}

\item{to.cart}{Convert to cartesian coordinates? (default \code{TRUE}).}
}
\description{
Workhorse: original range to cartesian coords according to \code{circ_par} settings.
}

