% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_par}
\alias{circ_par}
\title{Set and query circplot settings}
\usage{
circ_par(...)
}
\arguments{
\item{...}{Key values pairs seperated by comma or a list of key value pairs as
e.g. returned by the function itself.}
}
\description{
Set and query circplot settings
}
\examples{
circ_par()
circ_par(x.from = c(0,1))   # change input range from 0 to 1
}

