% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_par_init}
\alias{circ_par_init}
\title{Initialize circplot parameters}
\usage{
circ_par_init()
}
\description{
The values are saved in the field \code{circ} using the \code{options} function.
The following parameters can be set:
\enumerate{
  \item x.from Range of original data (default \code{[0,1]}).
  \item x.to Range data is mapped to, usually 0 to 2pi.
  \item y.from Range of original data (default \code{[0,1]}).
  \item y.to Range data is mapped to (default \code{[1,1.1]})
  \item n.seg Number of segments used when drawing lines.
  \item rings Border radi of rings (defaults is three rings \code{c(1.0, 1.1, 1.2, 1.3)})
}
}

