% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_plot}
\alias{circ_plot}
\title{Set up default plotting area}
\usage{
circ_plot(xlim = NULL, ylim = NULL, ...)
}
\description{
This is just a convenient wrapper for \code{plot} for pure convencience and
quick results.
}

