% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_quantiles}
\alias{circ_quantiles}
\title{Calculate quantiles from circular data.}
\usage{
circ_quantiles(x, probs = seq(0, 1, 0.25))
}
\arguments{
\item{x}{Original data.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}
}
\description{
Functions calculates quantiles using the \code{quantile} function
from the \code{circular} package.
}

