% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_text}
\alias{circ_text}
\title{Draw text}
\usage{
circ_text(label, x, y, xadj = 0.5, ...)
}
\arguments{
\item{label}{A label.}

\item{x,y}{Coordinates in original system}

\item{xadj}{x adjustement. One of \code{c(0, .5, 1)}.}

\item{...}{passed to \code{plotrix::arctext}.}
}
\description{
Has to be rewritten to include nice \code{facing} options (see \code{circlize::circos.text}).
Currently \code{plotrix::arctext} has a buch when setting \code{clockwise = FALSE}
and  using the \code{end} argument. Once this is fixed (mail to Jim Lemon on 2016-03-18)
\code{circ_text} can be improved. Especially nicefacing would be good.
}

