% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circplot-package.R
\docType{package}
\name{circplot}
\alias{circplot}
\alias{circplot-package}
\title{\pkg{circplot} draws several graphical elements in polar fashion}
\description{
\pkg{circplot} draws several graphical elements in polar fashion
}
\note{
Some functionality is inspired by the \code{circlize} package.
}
\author{
Mark Heckmann
}
\keyword{circplot}
\keyword{package}

