% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scan1_circ.R
\name{plot_scan1_circ}
\alias{plot_scan1_circ}
\title{plot a genome scan as a circle}
\usage{
plot_scan1_circ(
  x,
  map,
  lodcolumn = 1,
  chr = NULL,
  gap = NULL,
  rlim = c(5, 6),
  start_angle = pi,
  clockwise = TRUE,
  xlim = NULL,
  ylim = NULL,
  chr_labels = TRUE,
  lwd = 2,
  col = "slateblue",
  lty = 1,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{Output of \code{\link[qtl2:scan1]{qtl2::scan1()}}}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[qtl2:insert_pseudomarkers]{qtl2::insert_pseudomarkers()}}.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{chr}{Selected chromosomes to plot; a vector of character
strings.}

\item{gap}{Gap between chromosomes. The default is 1\% of the total genome length.}

\item{rlim}{radius limits}

\item{start_angle}{angle to start}

\item{clockwise}{If true, go clockwise}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{chr_labels}{Whether to add chromosome labels}

\item{lwd}{line width for LOD curves}

\item{col}{color for LOD curves}

\item{lty}{line type for LOD curves}

\item{add}{Add curves to previous plot (need care about matching \code{rlim} and \code{ylim})}

\item{...}{Additional graphics parameters}
}
\value{
No return value
}
\description{
Plot LOD curves for a genome scan as a circle
}
\examples{
library(qtl2)
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
map <- insert_pseudomarkers(iron$gmap, step=1)
probs <- calc_genoprob(iron, map, error_prob=0.002)
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

out <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar)
kinship <- calc_kinship(probs, "loco")
out_lmm <- scan1(probs, pheno, kinship, covar, Xcovar)

max_lod <- max(c(maxlod(out), maxlod(out_lmm)))
plot_scan1_circ(out, map, ylim=c(0, max_lod), rlim=c(5, 7))
plot_scan1_circ(out_lmm, map, col="violetred", lty=2, add=TRUE,
                ylim=c(0, max_lod), rlim=c(5, 7))
}
