% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{text_rot}
\alias{text_rot}
\title{Plot text rotated depending angle to x-axis}
\usage{
text_rot(x, y = NULL, label, ...)
}
\arguments{
\item{x}{numeric vectors of coordinates where the text
    \code{labels} should be written.  If the length of \code{x} and
    \code{y} differs, the shorter one is recycled.}

\item{...}{further \link{graphical parameters} (from \code{\link{par}}),
    such as \code{srt}, \code{family} and \code{xpd}.}
}
\description{
Plot text rotated depending angle to x-axis
}
\examples{
# text automatically rotated with respüect to x-axis

lim <- c(-1.5,1.5)
plot(NULL, asp=1, xlim=lim, ylim=lim)
plotrix::draw.circle(0, 0, 1)
s <- 1.05
phis <- seq(0, 2*pi, by=.4)
for (phi in phis) {
  v <- c(cos(phi), sin(phi))
  segments(0,0, v[1], v[2])
  text_rot(v * 1.1, label=phi)
}
}

