% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_angle}
\alias{vec_angle}
\title{Angle / radian between two vectors.}
\usage{
vec_angle(a, b, rad = TRUE, mode = 1)
}
\arguments{
\item{a,b}{Vector}

\item{rad}{Angular measure is radian?  (default \code{TRUE}).}

\item{mode}{1= smallest angle in [0, 180); 2=anti-clockwise angle with respect
vector b in [0, 360).}
}
\value{
Angle
}
\description{
Function comes in two flavors. Either it calculates the smallest angle
[0, 180) or the anti-clockwise angle with respect to the second
vector [0, 360).
}
\keyword{internal}

