% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_angles}
\alias{vec_angles}
\title{Angles / radian between two or more vectors.}
\usage{
vec_angles(a, b, rad = TRUE, mode = 1)
}
\arguments{
\item{a}{A matrix or a vector}

\item{b}{A reference vector to find angle against.}

\item{rad}{Angular measure is radian?  (default \code{TRUE}).}

\item{mode}{1= smallest angle in [0, 180); 2=anti-clockwise angle with respect
vector b in [0, 360).}
}
\value{
Vector of angles
}
\description{
Function comes in two flavors. Either it calculates the smallest angle
[0, 180) or the anti-clockwise angle with respect to the second
vector [0, 360).
}
\examples{
v <- c(1,0)   # point vector

rad <- seq(0, 2*pi, length=10)    # radians to rotate
m <- cbind(cos(rad), sin(rad))    # 2d m matrix 

# mode 1: between 0 and 180 degrees, minimal angle / radian
vec_angles(v, c(1,0))
vec_angles(m, c(1,0))

# mode 2: between 0 and 360 degrees, anti-clockwise 
vec_angles(v, c(1,0), mode=2)
vec_angles(m, c(1,0), mode=2)

}
\keyword{internal}

