% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2circ.R
\name{xy2circ}
\alias{xy2circ}
\title{convert cartesian coordinates to circular coordinates}
\usage{
xy2circ(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  rlim = c(1, 2),
  start_angle = pi,
  clockwise = TRUE
)
}
\arguments{
\item{x}{vector of x coordinates, or a 2-column matrix with columns x and y}

\item{y}{vector of y coordinates (ignored if x is a 2-column matrix)}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{rlim}{radius limits for plot}

\item{start_angle}{angle to start}

\item{clockwise}{If true, go clockwise}
}
\value{
2-column data frame of (x,y) locations in circle plot
}
\description{
convert standard cartesian coordiates to coordinates in a circular plot
}
\examples{
x <- seq(0, 18*pi, len=251)
y <- sin(x)
rlim <- c(4, 7)
xl <- c(-rlim[2], rlim[2])*1.1
yl <- range(y)
par(bty="n", pty="s")
plot(0, 0, type="n", xaxs="i", yaxs="i", xlim=xl, ylim=xl,
     xaxt="n", yaxt="n", xlab="", ylab="")
pos <- xy2circ(x, y, rlim=rlim)
lines(pos, lwd=2, col="slateblue")
circ <- xy2circ(x, rep(0, length(x)), rlim=rlim, ylim=yl)
lines(circ, lwd=2)

}
