% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{circ_quantiles}
\alias{circ_quantiles}
\title{Calculate \code{\link{fivenum}} avlues for boxplots from circular data.}
\usage{
circ_quantiles(x, probs = seq(0, 1, 0.25))
}
\description{
Functions calculates medians using the \code{quantile} function
from the \code{circular} package.
}

