% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circplot-package.R
\docType{package}
\name{circplot-package}
\alias{circplot}
\alias{circplot-package}
\title{circplot: Circular Plots}
\description{
Tools to make circular plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/circplot}
  \item Report bugs at \url{https://github.com/kbroman/circplot/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
