% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scan1_circ.R
\name{plot_scan1_circ}
\alias{plot_scan1_circ}
\title{plot a genome scan as a circle}
\usage{
plot_scan1_circ(
  x,
  map,
  lodcolumn = 1,
  chr = NULL,
  gap = NULL,
  rlim = c(5, 6),
  start_angle = pi,
  clockwise = TRUE,
  xlim = NULL,
  ylim = NULL,
  chr_labels = TRUE,
  ...
)
}
\arguments{
\item{x}{Output of \code{\link[qtl2:scan1]{qtl2::scan1()}}}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[qtl2:insert_pseudomarkers]{qtl2::insert_pseudomarkers()}}.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{chr}{Selected chromosomes to plot; a vector of character
strings.}

\item{gap}{Gap between chromosomes. The default is 1\% of the total genome length.}

\item{rlim}{radius limits}

\item{start_angle}{angle to start}

\item{clockwise}{If true, go clockwise}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{chr_labels}{Whether to add chromosome labels}

\item{...}{Additional graphics parameters}
}
\value{
No return value
}
\description{
Plot LOD curves for a genome scan as a circle
}
